requirements.txt
python-dotenv==1.0.0
web3==6.11.3
